% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amgut.R
\docType{data}
\name{combinedamgut}
\alias{combinedamgut}
\title{A Subdata from the American Gut Project study data}
\format{
combinedamgut
}
\usage{
data(combinedamgut)
}
\description{
A pre-processed OTU table and clinical data were obtained from the American
 Gut Project, available in the SpiecEasi R package.
}
\examples{
data(combinedamgut)
}
\references{
McDonald D, et al. American Gut: an Open Platform for Citizen Science Microbiome Research. mSystems, 2018;3(3):e00031-18.
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5954204/}{PubMed})
}
\keyword{1}
\keyword{An}
\keyword{SOHPIE}
\keyword{data}
\keyword{example}
\keyword{for}
\keyword{no.}
\keyword{package.}
