% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietswap.R
\docType{data}
\name{combineddietswap}
\alias{combineddietswap}
\title{A Subdata from the Diet Swap study data}
\format{
`combineddietswap`
}
\usage{
data(combineddietswap)
}
\description{
A pre-processed OTU table and clinical data from the geographical epidemiology study
 (aka the Diet Exchange Study) is available in the microbiome R package.
}
\examples{
data(combineddietswap)
}
\references{
O'Keefe, SJ, et al. Fat, fibre and cancer risk in african americans and
rural africans. Nat Commun. 2015;6:6342.
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4415091/}{PubMed})
}
\keyword{1}
\keyword{An}
\keyword{SOHPIE}
\keyword{data}
\keyword{example}
\keyword{for}
\keyword{no.}
\keyword{package.}
